+++ backend/core/src/test/kotlin/nl/nlportal/core/util/MapperTest.kt
@@ -1,0 +1,32 @@
+/*
+ * Copyright 2026 Ritense BV, the Netherlands.
+ *
+ * Licensed under EUPL, Version 1.2 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * https://joinup.ec.europa.eu/collection/eupl/eupl-text-eupl-12
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" basis,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package nl.nlportal.core.util
+
+import java.time.LocalDateTime
+import org.junit.jupiter.api.Assertions.assertEquals
+import org.junit.jupiter.api.Test
+import org.junit.jupiter.api.TestInstance
+
+@TestInstance(TestInstance.Lifecycle.PER_CLASS)
+class MapperTest {
+    @Test
+    fun `serialize LocalDateTime`() {
+        val localDateTimeTest = LocalDateTime.of(2021, 1, 1, 1, 0, 0)
+        val result = Mapper.get().writeValueAsString(localDateTimeTest)
+
+        assertEquals("\"2021-01-01T01:00:00.000Z\"", result)
+    }
+}